feat : mock student connect

+++ src/main/java/com/geupjo/koreantiger/dto/response/BiWeeklyAchievementResponseDto.java
@@ -61,7 +61,8 @@ public record BiWeeklyAchievementResponseDto(
             if (thisWeek.totalLearningTime == 0) {
                 return String.valueOf(0);
             }
-            double rate = (double) thisWeek.totalLearningTime() / (double) lastWeek.totalLearningTime();
+
+            double rate = (double) (thisWeek.totalLearningTime() / lastWeek.totalLearningTime());
 
             return String.format("%.1f", rate);
         }
